learn-awscdk-with-claude/simple-ecs-app/requirements.txt
# AWS CDK Core Libraries
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# AWS SDK for Python (boto3)
boto3>=1.26.0

# 開発時に便利なツール（オプション）
# pytest>=7.0.0  # テスト用
# black>=22.0.0   # コードフォーマッター用